+++ packages/leancode_cubit_utils/test/request_cubit_builder_test.dart
@@ -159,7 +159,7 @@ void main() {
       await tester.pumpAndSettle();
     });
 
-    testWidgets('shows custom onRefresh widget when onRefresh is provided', (
+    testWidgets('shows custom refresh widget when onRefresh is provided', (
       tester,
     ) async {
       final cubit = FakeRequestCubit.refresh();
@@ -168,7 +168,7 @@ void main() {
         TestPage(
           child: RequestCubitBuilder(
             cubit: cubit,
-            onRefresh: (context, data) => const Text('Refreshing'),
+            onRefreshing: (context, data) => const Text('Refreshing'),
             onSuccess: (context, data) => const Text('Success'),
           ),
         ),
@@ -178,7 +178,7 @@ void main() {
       expect(find.text('Refreshing'), findsOneWidget);
     });
 
-    testWidgets('onRefresh not provided does not cause an error', (
+    testWidgets('onRefreshing not provided does not cause an error', (
       tester,
     ) async {
       final cubit = FakeRequestCubit.refresh();